requirements.txt
aiometer==0.3.0
anyio==3.6.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
executing==1.1.1
filelock==3.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.16.2
ipython==8.5.0
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.1
jmespath==1.0.1
jupyter-client==7.4.4
jupyter-core==4.11.2
kucoin-python==1.0.11
lxml==4.9.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
Protego==0.2.1
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
pyzmq==24.0.1
queuelib==1.6.2
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
stack-data==0.5.1
tldextract==3.4.0
tornado==6.2
traitlets==5.5.0
Twisted==22.8.0
typing-extensions==4.4.0
urllib3==1.26.12
w3lib==2.0.1
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
zope.interface==5.5.0
